# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List the user's programs
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Generate a personalized program with AI
      description: Uses the self-assessment, chosen path, and duration to generate a full plan with levels and daily tasks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created program with generated levels and tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /programs/current:
    get:
      operationId: getCurrentProgram
      tags: [programs]
      summary: Get the user's active program with levels and tasks
      responses:
        "200":
          description: Current program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
        "404":
          description: No active program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get a program with its levels and tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteProgram
      tags: [programs]
      summary: Delete a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /programs/{id}/activate:
    post:
      operationId: activateProgram
      tags: [programs]
      summary: Make a program the active one
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
...
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List reminders
      responses:
        "200":
          description: Reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chat/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: Get the AI coach chat history
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the AI coach and get a reply
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Assistant reply plus the stored user message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatReply"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: Bearer token for mobile clients. Web uses the session cookie.
    User:
      type: object
      required: [id, email, name, xp, level, streak, hasProgram, onboarded]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        xp:
          type: integer
        level:
          type: integer
          description: Overall account level derived from XP
        streak:
          type: integer
          description: Current daily streak in days
        hasProgram:
          type: boolean
        onboarded:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    ProgramInput:
      type: object
      required: [pathKey, durationDays]
      properties:
        pathKey:
          type: string
        durationDays:
          type: integer
          description: 30 or 60
        about:
          type: string
        currentSituation:
          type: string
        biggestChallenge:
          type: string
        motivation:
          type: string
    Program:
      type: object
      required: [id, pathKey, pathTitle, durationDays, status, currentLevel, totalLevels, progressPct, createdAt]
      properties:
        id:
          type: integer
        pathKey:
          type: string
        pathTitle:
          type: string
        title:
          type: string
        durationDays:
          type: integer
        status:
          type: string
          description: active, completed, or archived
        currentLevel:
          type: integer
        totalLevels:
          type: integer
        progressPct:
          type: integer
        accent:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProgramDetail:
      type: object
      required: [id, pathKey, pathTitle, durationDays, status, currentLevel, totalLevels, progressPct, levels, createdAt]
      properties:
        id:
          type: integer
        pathKey:
          type: string
        pathTitle:
          type: string
        title:
          type: string
        summary:
          type: ["string", "null"]
        durationDays:
          type: integer
        status:
          type: string
        currentLevel:
          type: integer
        totalLevels:
          type: integer
        progressPct:
          type: integer
        accent:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        levels:
          type: array
          items:
            $ref: "#/components/schemas/Level"
    Level:
      type: object
      required: [id, levelNumber, title, status, xpReward, tasks]
      properties:
        id:
          type: integer
        levelNumber:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          description: locked, active, or completed
        xpReward:
          type: integer
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
    Task:
      type: object
      required: [id, levelId, levelNumber, dayNumber, title, completed, xp]
      properties:
        id:
          type: integer
        levelId:
          type: integer
        levelNumber:
          type: integer
        dayNumber:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        timeOfDay:
          type: ["string", "null"]
          description: Suggested time like "07:00" or "Morning"
        durationMinutes:
          type: ["integer", "null"]
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        xp:
          type: integer
...
    Reminder:
      type: object
      required: [id, title, timeOfDay, enabled, daysOfWeek]
      properties:
        id:
          type: integer
        title:
          type: string
        timeOfDay:
          type: string
          description: "HH:MM 24-hour"
        enabled:
          type: boolean
        daysOfWeek:
          type: array
          description: 0=Sunday .. 6=Saturday
          items:
            type: integer
    ReminderInput:
      type: object
      required: [title, timeOfDay]
      properties:
        title:
          type: string
        timeOfDay:
          type: string
        enabled:
          type: boolean
        daysOfWeek:
          type: array
          items:
            type: integer
    ReminderUpdate:
      type: object
      properties:
        title:
          type: string
        timeOfDay:
          type: string
        enabled:
          type: boolean
        daysOfWeek:
          type: array
          items:
            type: integer
    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
          description: user or assistant
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    ChatReply:
      type: object
      required: [userMessage, assistantMessage]
      properties:
        userMessage:
          $ref: "#/components/schemas/ChatMessage"
        assistantMessage:
          $ref: "#/components/schemas/ChatMessage"