# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessage {
  id: number;
  /** user or assistant */
  role: string;
  content: string;
  createdAt: Date;
}