# lib/api-zod/src/generated/types/chatReply.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatReply {
  userMessage: ChatMessage;
  assistantMessage: ChatMessage;
}