# lib/api-zod/src/generated/types/reminderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReminderInput {
  title: string;
  timeOfDay: string;
  enabled?: boolean;
  daysOfWeek?: number[];
}