# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  levelId: number;
  levelNumber: number;
  dayNumber: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /**
     * Suggested time like "07:00" or "Morning"
     * @nullable
     */
  timeOfDay?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  completed: boolean;
  /** @nullable */
  completedAt?: Date | null;
  xp: number;
}