# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  xp: number;
  /** Overall account level derived from XP */
  level: number;
  /** Current daily streak in days */
  streak: number;
  hasProgram: boolean;
  onboarded: boolean;
  createdAt?: Date;
}